Added Construct_initial_points_from_labeled_image The example use this struct initialize_triangulation_from_labeled_image also uses it Renamed parameter initial_points to construct_initial_points
