+++ STL_Extension/include/CGAL/STL_Extension/internal/parameters_interface.h
@@ -335,7 +335,7 @@ CGAL_add_named_parameter_with_compatibility(surface_patch_index_t, surface_patch
 CGAL_add_named_parameter_with_compatibility_ref_only(weights_param_t, weights_param, weights)
 CGAL_add_named_parameter_with_compatibility(features_detector_param_t, features_detector_param, features_detector)
 CGAL_add_named_parameter_with_compatibility(input_features_param_t, input_features_param, input_features)
-CGAL_add_named_parameter_with_compatibility(initial_points_param_t, initial_points_param, initial_points)
+CGAL_add_named_parameter_with_compatibility(initial_points_param_t, initial_points_param, construct_initial_points)
 
 CGAL_add_named_parameter_with_compatibility(edge_size_param_t, edge_size_param, edge_size)
 CGAL_add_named_parameter_with_compatibility_cref_only(edge_sizing_field_param_t, edge_sizing_field_param, edge_sizing_field)
